# docs/openapi/core.openapi.yaml
openapi: 3.0.3
info:
  title: Pulse 109 Core API
  version: 1.0.0
  description: >-
    Public application contract. Core API validates access and orchestrates
    PostgreSQL, Qdrant and the internal ML service. It never represents a
    write to the official 109 system. Demo mode accepts an optional x-pulse-role
    selector; non-demo mode requires identity headers from a trusted gateway.
servers:
  - url: http://localhost:8080
    description: Local Nginx gateway
...
paths:
...
  /api/v1/forecast:
    get:
      tags: [forecast]
      summary: Forecast 30, 60 and 90 days
      security: [{ pulseRole: [] }, {}]
      parameters:
        - name: horizon
          in: query
          schema: { type: integer, enum: [30, 60, 90] }
        - $ref: '#/components/parameters/RegionIdQuery'
        - name: topic_id
          in: query
          schema: { type: string }
        - name: service_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: district
          in: query
          schema: { type: string }
        - name: channel
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Forecast demand and expected peaks; capacity is explicitly unavailable without verified operational inputs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ForecastResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/models:
    get:
      tags: [models]
      summary: List model versions and statuses
      description: ML_REVIEWER or ADMIN only.
      security: [{ pulseRole: [] }, {}]
      responses:
        '200':
          description: Model registry entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ModelVersion' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/models/{model_id}:
    get:
      tags: [models]
      summary: Get a model version
      description: ML_REVIEWER or ADMIN only.
      parameters:
        - name: model_id
          in: path
          required: true
          schema: { type: string }
      security: [{ pulseRole: [] }, {}]
      responses:
        '200': { description: Request completed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    pulseRole:
      type: apiKey
      in: header
      name: x-pulse-role
      description: Demo role selector. Production identity comes from a trusted gateway.
  parameters:
...
    RegionIdQuery:
      name: region_id
      in: query
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Forbidden:
      description: Role is not allowed
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    ForecastResponse:
      type: object
      required: [source, model_version, model, status, insufficient_history, horizon_days, history, forecast_start, points, expected_peaks, backtest, capacity_assessment, manager_signals]
      properties:
        source: { type: string }
        model_version: { type: string, description: Forecast provider version; never the embedding version. }
        model: { type: string }
        status: { type: string, enum: [OK, INSUFFICIENT_HISTORY, DEMO_ONLY] }
        insufficient_history: { type: boolean }
        horizon_days: { type: integer }
        history: { type: array, items: { $ref: '#/components/schemas/TimeSeriesPoint' } }
        forecast_start: { type: string, format: date, nullable: true, description: First forecast date, or null when history is insufficient. }
        points: { type: array, items: { $ref: '#/components/schemas/TimeSeriesPoint' } }
        expected_peaks: { type: array, items: { type: string } }
        backtest: { $ref: '#/components/schemas/ForecastBacktest' }
        capacity_assessment: { $ref: '#/components/schemas/ForecastCapacityAssessment' }
        run_id: { type: string, nullable: true }
        issued_at: { type: string, format: date-time, nullable: true }
        reforecast: { allOf: [{ $ref: '#/components/schemas/ForecastReforecast' }], nullable: true }
        manager_signals: { type: array, items: { $ref: '#/components/schemas/ForecastManagerSignal' } }
    ForecastCapacityAssessment:
      type: object
      description: Capacity is not calculated until real staffing, handling-time or throughput, schedule and service-level/SLA inputs are available.
      required: [status, missing_inputs]
      properties:
        status: { type: string, enum: [DATA_UNAVAILABLE] }
        missing_inputs:
          type: array
          uniqueItems: true
          items: { type: string, enum: [STAFFING, HANDLING_TIME_OR_THROUGHPUT, SCHEDULE, SERVICE_LEVEL_TARGET_OR_SLA] }
...
    ForecastReforecast:
      type: object
      required: [previous_run_id, previous_model_version, previous_issued_at, actual_observations, future_comparisons, peak_change]
      properties:
        previous_run_id: { type: string }
        previous_model_version: { type: string }
        previous_issued_at: { type: string, format: date-time }
        actual_observations: { type: array, items: { $ref: '#/components/schemas/ForecastActualObservation' } }
        future_comparisons: { type: array, items: { $ref: '#/components/schemas/ForecastFutureComparison' } }
        peak_change: { allOf: [{ $ref: '#/components/schemas/ForecastPeakChange' }], nullable: true }
    ForecastActualObservation:
      type: object
      required: [date, previous_forecast, actual, absolute_error]
      properties:
        date: { type: string, format: date }
        previous_forecast: { type: integer, minimum: 0 }
        actual: { type: integer, minimum: 0 }
        absolute_error: { type: integer, minimum: 0 }
    ForecastFutureComparison:
      type: object
      required: [date, previous_forecast, updated_forecast, delta]
      properties:
        date: { type: string, format: date }
        previous_forecast: { type: integer, minimum: 0 }
        updated_forecast: { type: integer, minimum: 0 }
        delta: { type: integer }
    ForecastPeakChange:
      type: object
      required: [previous_peak_date, updated_peak_date, previous_peak, updated_peak, delta, policy_version, threshold, status, manager_signal_id]
      properties:
        previous_peak_date: { type: string, format: date }
        updated_peak_date: { type: string, format: date }
        previous_peak: { type: integer, minimum: 0 }
        updated_peak: { type: integer, minimum: 0 }
        delta: { type: integer }
        policy_version: { type: string }
        threshold: { type: number, nullable: true, minimum: 0, description: Maximum of the previous and updated rolling backtest MAE. }
        status: { type: string, enum: [SIGNIFICANT, WITHIN_BACKTEST_ERROR, BACKTEST_UNAVAILABLE, SOURCE_NOT_VERIFIED_REAL, SIGNAL_CREATED] }
        manager_signal_id: { type: string, nullable: true }
    ForecastManagerSignal:
      type: object
      required: [id, run_id, previous_run_id, created_at, previous_peak, updated_peak, delta, threshold, policy_version]
      properties:
        id: { type: string }
        run_id: { type: string }
        previous_run_id: { type: string }
        created_at: { type: string, format: date-time }
        previous_peak: { type: integer, minimum: 0 }
        updated_peak: { type: integer, minimum: 0 }
        delta: { type: integer }
        threshold: { type: number, minimum: 0 }
        policy_version: { type: string }
    ForecastBacktest:
      type: object
      properties:
        status: { type: string, enum: [OK, INSUFFICIENT_HISTORY, DEMO_ONLY] }
        history_days: { type: integer, minimum: 0 }
        observed_days: { type: integer, minimum: 0 }
        required_days: { type: integer, minimum: 1 }
        sample_count: { type: integer, minimum: 0 }
        window_count: { type: integer, minimum: 0 }
        mae: { type: number, nullable: true, minimum: 0 }
        rmse: { type: number, nullable: true, minimum: 0 }
        wape: { type: number, nullable: true, minimum: 0 }
        smape: { type: number, nullable: true, minimum: 0 }
    TimeSeriesPoint:
      type: object
      required: [date, tickets, resolved]
      properties:
        date: { type: string, format: date }
        tickets: { type: integer }
        resolved: { type: integer }
...
    ModelVersion:
      type: object
      required: [model_version, model_family, status]
      properties:
        model_version: { type: string }
        model_family: { type: string, enum: [classifier, embedder, forecast] }
        dataset_version: { type: string }
        artifact_checksum: { type: string }
        status: { type: string, enum: [candidate, production, rejected, archived] }
...